scripts/requirements.txt
# FalconPy - CrowdStrike Falcon SDK for Python
crowdstrike-falconpy>=1.4.0
